requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
cryptg==0.4.0
pyyaml==6.0.2
pydantic==2.9.2
ruff==0.6.9